requirements.txt
anyascii==0.2.0
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.44.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-modelcluster==5.1
django-split-settings==1.1.0
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.7.2
draftjs-exporter==2.1.7
drf-spectacular==0.17.3
et-xmlfile==1.1.0
html5lib==1.1
idna==3.2
inflection==0.5.1
jsonschema==3.2.0
l18n==2020.6.1
oauthlib==3.1.1
openpyxl==3.0.7
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
sentry-sdk==1.3.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
wagtail==2.14.1
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
